Refactor db operations to different modules
