+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/index.js",
   "scripts": {
     "start": "node src/index.js",
-    "init-db": "node src/db/create-tables.js",
+    "init-db": "node src/db/main.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
